Moved filters to filters.py submodule
